pyproject.toml
[tool.poetry]
name = "sendcloud"
version = "1.0.0"
description = "Sendcloud feed"
readme = "README.md"
authors = ["Pranav Singh <pranav.singh@gmail.com>"]

[tool.black]
line-length = 80
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.hg
  | \.venv
  | \.mypy_cache
  | _build
  | build
  | dist
  | env
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 80
skip = ["migrations"]

[tool.pytest.ini_options]
python_files =["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*", "*TestCase"]
addopts = "--no-migrations --reuse-db --ds=sendcloud.settings_test"

[tool.poetry.dependencies]
python = "^3.7"
celery = "^4.4.7"
django-celery-beat = "^2.1.0"
django-enumfield = "^1.6.0"
django-environ = "^0.4.5"
django = "^2.2.0"
djangorestframework = "^3.11.0"
model-bakery = "^1.1.0"
mysqlclient = "^2.0.1"
redis = "^3.5.3"
requests = "^2.22.0"
responses = "^0.10.9"
uWSGI = "^2.0.19.1"
beautifulsoup4 = "^4.10.0"
lxml = "^4.6.3"


[tool.poetry.dev-dependencies]
ipdb = "^0.13.3"
coverage = "^5.2.1"
pre-commit = "^2.7.1"
flake8 = "^3.8.3"
pytest = "^6.0.1"
pytest-django = "^3.9.0"
isort = {extras = ["pyproject"], version = "^5.5.4"}
black = "^20.8b1"
pytest-mock = "^3.3.1"
celery-mock = "^1.1.7"
pdbpp = "^0.10.2"
